feat: #48 내 펀딩 리스트 서버 연동

+++ Android/FundYou/domain/src/main/java/com/ssafy/fundyou/domain/model/funding/FundingInfoModel.kt
@@ -8,5 +8,5 @@ data class FundingInfoModel(
     val endDate: Long,
     val currentFundingPrice: Int,
     val totalPrice: Int,
-    val percentage: Float,
+    val percentage: Int,
 )
